requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
motor==3.4.0
pymongo==4.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
sse-starlette==2.1.0
deepgram-sdk==3.2.7
groq==0.9.0
google-generativeai==0.7.2
python-multipart==0.0.9
